chore: move enums to their own directory
