+++ ProjectAlchemy.Core/Domain/Project.cs
@@ -1,3 +1,4 @@
+using ProjectAlchemy.Core.Enums;
 using ProjectAlchemy.Core.Exceptions;
 using ProjectAlchemy.Core.Helpers;
