+++ ProjectAlchemy.Tests/Integration/UserServiceTests.cs
@@ -1,6 +1,7 @@
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Enums;
 using ProjectAlchemy.Core.Exceptions;
 using ProjectAlchemy.Core.Services;
 using ProjectAlchemy.Persistence;
