+++ ProjectAlchemy.Web/Dtos/CreateIssueRequest.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Enums;
 
 namespace ProjectAlchemy.Web.Dtos;
