+++ ProjectAlchemy.Web/Dtos/IssuePatch.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Enums;
 
 namespace ProjectAlchemy.Web.Dtos;
